// lib/types/settingTypes.ts
export interface UserSettings {
    id?: string; // UUID or unique identifier
    address?: string;
    phone_number?: string;
    email?: string;
    linkedin_url?: string;
    github_url?: string;
    website_1_url?: string;
    website_2_url?: string;
    about_me?: string;
    about_short?: string;
  }